backend/requirements.txt
flask==2.0.1
werkzeug==2.0.3
flask-cors==3.0.10
flask-sqlalchemy==2.5.1
sqlalchemy==1.4.46
psycopg2-binary==2.9.1
scikit-learn==1.0.2
pandas==1.4.0
numpy==1.22.1
requests==2.27.1
beautifulsoup4==4.10.0
tldextract==3.1.2
joblib==1.1.0
python-dotenv==0.19.2
gunicorn==20.1.0